requirements.txt
yfinance
pandas
mysql-connector-python
schedule
streamlit
plotly
pdoc